requirements.txt
streamlit>=1.22.0
Pillow>=9.0.0 